refactor: move github config and attempts into service boundary
