+++ api/app/src/__tests__/github-bind-attempts.test.ts
@@ -24,7 +24,7 @@ const {
   issueGitHubOAuthAttempt,
   lookupGitHubInstallAttempt,
   lookupGitHubOAuthAttempt,
-} = await import("../github/bind-attempts");
+} = await import("../services/github/setup/attempts");
 
 beforeEach(() => {
   redisSetMock.mockReset();
